# packages/api-client/lib/model/authResponse.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * notebook-service
 * OpenAPI spec version: 1.0.0
 */
import type { UserDTO } from './userDTO';

export interface AuthResponse {
  token: string;
  user: UserDTO;
}