# packages/api-client/lib/model/createNotebookRequest.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * notebook-service
 * OpenAPI spec version: 1.0.0
 */

export interface CreateNotebookRequest {
  name: string;
  workspaceId: number;
}