# packages/api-client/lib/model/createPageRequest.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * notebook-service
 * OpenAPI spec version: 1.0.0
 */

export interface CreatePageRequest {
  title: string;
  content?: string;
  notebookId: number;
}