# packages/api-client/lib/model/notebookDTO.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * notebook-service
 * OpenAPI spec version: 1.0.0
 */

export interface NotebookDTO {
  id: number;
  name: string;
  workspaceId: number;
}